app/schemas/common.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict


class ORMModel(BaseModel):
    model_config = ConfigDict(from_attributes=True)


class HealthResponse(BaseModel):
    status: str


class Timestamped(ORMModel):
    id: int
    created_at: datetime
